fix(HTD-90): handle linter errors
